fix: tipo de dato

+++ 04-teslo-shop/src/product/entities/product.entity.ts
@@ -15,7 +15,7 @@ export class Product {
   })
   title: string;
 
-  @Column('numeric', {
+  @Column('float', {
     default: 0
   })
   price: number;
